+++ include/global.h
@@ -995,7 +995,7 @@ struct SaveBlock1
     /*0x238*/ struct Pokemon playerParty[PARTY_SIZE];
     /*0x490*/ u32 money;
     /*0x494*/ u16 coins;
-    /*0x496*/ u16 registeredItem; // registered for use with SELECT button
+    /*0x496*/ u16 registeredItems[REGISTERED_ITEMS_COUNT];
     /*0x560*/ struct ItemSlot bagPocket_Items[BAG_ITEMS_COUNT];
     /*0x560*/ struct ItemSlot bagPocket_Medicine[BAG_MEDICINE_COUNT];
     /*0x560*/ struct ItemSlot bagPocket_BattleItems[BAG_BATTLE_ITEMS_COUNT];