+++ src/new_game.c
@@ -45,6 +45,7 @@
 #include "mystery_gift.h"
 #include "union_room_chat.h"
 #include "constants/items.h"
+#include "registered_items.h"
 
 extern const u8 EventScript_ResetAllMapFlags[];
 
@@ -183,7 +184,7 @@ void NewGameInitData(void)
     ResetPokemonStorageSystem();
     ClearRoamerData();
     ClearRoamerLocationData();
-    gSaveBlock1Ptr->registeredItem = ITEM_NONE;
+    ClearItemsRegistered();
     ClearBag();
     NewGameInitItems();
     ClearPokeblocks();