+++ src/party_menu.c
@@ -55,7 +55,7 @@
 #include "script.h"
 #include "sound.h"
 #include "sprite.h"
-#include "start_menu.h"
+#include "sub_menus.h"
 #include "string_util.h"
 #include "strings.h"
 #include "task.h"